fix/ #65 api 수정
