--- backend/src/main/java/com/devnear/web/dto/project/ProposalInquiryResponse.java
@@ -1,7 +1,0 @@
-package com.devnear.web.dto.project;
-
-/**
- * 프로젝트 제안 "문의하기" 시 반환되는 채팅방 식별자.
- */
-public record ProposalInquiryResponse(Long chatRoomId) {
-}
