+++ backend/src/main/java/com/devnear/web/dto/proposal/ProposalInquiryResponse.java
@@ -1,0 +1,4 @@
+package com.devnear.web.dto.proposal;
+
+public record ProposalInquiryResponse(Long chatRoomId) {
+}
